// backend/src/domain/enums/ReturnExchangeRequestStatus.ts
export enum ReturnExchangeRequestStatus {
    EXCHANGE_REQUESTED = 'EXCHANGE_REQUESTED',         // TROCA SOLICITADA
    EXCHANGE_ACCEPTED = 'EXCHANGE_ACCEPTED',           // TROCA ACEITA
    EXCHANGE_REJECTED = 'EXCHANGE_REJECTED',           // TROCA RECUSADA
    EXCHANGE_COMPLETED = 'EXCHANGE_COMPLETED',         // TROCA CONCLUÍDA

    RETURN_REQUESTED = 'RETURN_REQUESTED',             // DEVOLUÇÃO SOLICITADA
    RETURN_REJECTED = 'RETURN_REJECTED',               // DEVOLUÇÃO RECUSADA
    RETURN_COMPLETED = 'RETURN_COMPLETED'              // DEVOLUÇÃO CONCLUÍDA
}